fix(lint): resolve CGO_ENABLED=0 golangci-lint failures (CI mode)
CI runs golangci-lint with CGO_ENABLED=0, which includes the //go:build !cgo
pkg/gateway and tests/perf files that a default (CGO=1) local run excludes —
surfacing two issues not caught locally:

- nolintlint: the //nolint:unused on tests/perf/percentile_test.go's
  logLatencyDistribution was redundant under CGO=0 (its !cgo caller is compiled,
  so the func is used). Moved logLatencyDistribution into a new //go:build !cgo
  file (tests/perf/latency_log_test.go) so it is compiled — and used — only where
  its caller exists; dropped the now-unused "testing" import from percentile_test.go.
  computePercentile stays untagged (also used by the untagged benchmark_media_test).
- dupl: TestChannelRouting_PutNullRemovesBinding and _PutEmptyStringTreatedAsRemove
  were near-identical; extracted assertRemoveBindingViaPut(t, removeBody) helper.

Verified clean under both CGO_ENABLED=0 and =1: golangci-lint 0 issues, gofmt 0,
build + routing/perf tests pass.
